MTDomeTrajectory: fix shutter vignetting code and test it
Fix shutter config field names.
Update code comments to explain that the "shutter" vignetting status
will be correct, if shutter position is reported,
but is not yet used as part of the overall "vignetted" status
in the telescopeVignetted event.

This is because the MTDome CSC cannot report shutter position
yet, when run with the real dome controller, but it does report
shutter position when run in simulation mode.